add some casts
git-svn-id: https://svn.r-project.org/R/trunk@64167 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/extra/intl/langprefs.c
@@ -55,7 +55,7 @@ const char *_nl_language_preferences_default(void)
             if (preferences != NULL && CFGetTypeID(preferences) == CFArrayGetTypeID())
             {
                 CFArrayRef prefArray = (CFArrayRef)preferences;
-                int n = CFArrayGetCount(prefArray);
+                long n = CFArrayGetCount(prefArray);
                 char buf[256];
                 size_t size = 0;
                 int i;
